% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyscreenshot-package.R
\docType{package}
\name{shinyscreenshot-package}
\alias{shinyscreenshot}
\alias{shinyscreenshot-package}
\title{shinyscreenshot: Capture Screenshots of Entire Pages or Parts of Pages in 'Shiny'}
\description{
Capture screenshots in 'Shiny' applications. Screenshots can either be of the entire viewable page, or a specific section of the page. The captured image is automatically downloaded as a PNG image, or it can also be saved on the server. Powered by the 'html2canvas' JavaScript library.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/daattali/shinyscreenshot}
  \item \url{https://daattali.com/shiny/shinyscreenshot-demo/}
  \item Report bugs at \url{https://github.com/daattali/shinyscreenshot/issues}
}

}
\author{
\strong{Maintainer}: Dean Attali \email{daattali@gmail.com} (\href{https://orcid.org/0000-0002-5645-3493}{ORCID})

Authors:
\itemize{
  \item Niklas von Hertzen \email{niklasvh@gmail.com} (html2canvas library)
  \item Eli Grey \email{~@eligrey.com} (FileSaver library)
}

}
\keyword{internal}
